Code Cleanup
More like delete everything and start  over, but with a focus of documenting and trying to make it a bit more readable.

I have it almost back to full working game. It needs to check the current turn and have it change the textbox. Other then that, its about there.